Refactor Node Graph types into messages
This commit moves the language server types from inside the node graph provider
into the generic area messages.
